pyproject.toml
[project]
name = "monitoring"
version = "0.1.0"
description = "Application monitoring and observability tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "requests>=2.32.4",
    "playwright>=1.40.0",
    "docker>=7.0.0",
    "apscheduler>=3.10.4",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "python-dateutil>=2.8.2",
    "python-telegram-bot>=22.3",
    "paramiko>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
